refactor(board): show sidebar in task board page
